+++ R/github-actions.R
@@ -4,51 +4,44 @@
 #' GitHub using [GitHub Actions](https://github.com/features/actions). These functions
 #' - Add the necessary configuration files and place them in `.Rbuildignore`.
 #' - Provide the markdown to insert a badge into your README
-#' @name actions
+#' @name github_actions
 NULL
 
 #' @section `use_github_actions()`:
-#' Adds a basic ‘R.yaml’ file to the `.github/workflows` directory of a
+#' Adds a basic ‘R-CMD-check.yaml’ file to the `.github/workflows` directory of a
 #'  package. This is a configuration file for the [GitHub
 #'  Actions](https://github.com/features/actions) service.
-#' @rdname actions
+#' @rdname github_actions
 #' @export
 use_github_actions <- function() {
   check_uses_github()
 
-  new <- use_action_check_release()
-
-  if (!new) {
-    return(invisible(new))
-  }
-  use_github_actions_badge("R-CMD-check")
-
-  invisible(new)
+  use_github_action_check_release()
 }
 
-#' @section `use_tidy_actions()`:
+#' @section `use_github_actions_tidy()`:
 #' Sets up tidyverse actions that check the R versions officially supported by
 #'   the tidyverse, (current release, devel and four previous versions). It also
 #'   adds two commands to be used in pull requests, `\document` to run
 #'   `roxygen2::roxygenise()` and update the PR, and `\style` to run
 #'   `styler::style_pkg()` and update the PR.
-#' @rdname actions
+#' @rdname github_actions
 #' @export
-use_tidy_actions <- function() {
+use_github_actions_tidy <- function() {
   check_uses_github()
 
-  new <- use_action_check_full() && use_action_pr_commands()
+  new <- use_github_action_check_full() && use_github_action_pr_commands()
 
   invisible(new)
 }
 
 #' @section `use_github_actions_badge()`:
 #' Only adds the [GitHub Actions](https://github.com/features/actions) badge. Use for a project
-#'   where GitHub actions is already configured.
+#'   where GitHub Actions is already configured.
 #' @param name The name to give to the [GitHub
 #'   Actions](https://github.com/features/actions) workflow.
 #' @export
-#' @rdname actions
+#' @rdname github_actions
 use_github_actions_badge <- function(name = "R-CMD-check") {
   check_uses_github()
 
@@ -59,14 +52,14 @@ use_github_actions_badge <- function(name = "R-CMD-check") {
   use_badge("R build status", url, img)
 }
 
-#' @section `use_action()`:
+#' @section `use_github_action()`:
 #' Use a specific action, either one of the example actions from
 #'   [r-lib/actions/examples](https://github.com/r-lib/actions/tree/master/examples) or a custom action
 #'   given by the `url` parameter.
-#' @param url The full URL to the action yaml file.
+#' @param url The full URL to the GitHub Actions yaml file.
 #' @inheritParams use_template
-#' @rdname actions
-use_action <- function(name,
+#' @rdname github_actions
+use_github_action <- function(name,
                        url = glue("https://raw.githubusercontent.com/r-lib/actions/master/examples/{name}"),
                        save_as = basename(url),
                        ignore = TRUE,
@@ -89,40 +82,58 @@ use_action <- function(name,
   invisible(new)
 }
 
-#' @section `use_action_check_release()`:
-#' This action installs the latest release R version on macOS and runs R CMD
-#'   check via the [rcmdcheck](https://github.com/r-lib/rcmdcheck) package.
-#' @rdname actions
+#' @section `use_github_action_check_release()`:
+#' This action installs the latest release R version on macOS and runs `R CMD check`
+#'   via the [rcmdcheck](https://github.com/r-lib/rcmdcheck) package.
+#' @rdname github_actions
 #' @export
-use_action_check_release <- function(save_as = "R.yaml", ignore = TRUE, open = FALSE) {
-  use_action("check-release.yaml", save_as = save_as, ignore = ignore, open = open)
+use_github_action_check_release <- function(save_as = "R-CMD-check.yaml", ignore = TRUE, open = FALSE) {
+  use_github_action("check-release.yaml", save_as = save_as, ignore = ignore, open = open)
 
   use_github_actions_badge("R-CMD-check")
 }
 
-#' @section `use_action_check_full()`:
+#' @section `use_github_action_check_full()`:
 #' This action installs the last 5 minor R versions and runs R CMD check
 #'   via the [rcmdcheck](https://github.com/r-lib/rcmdcheck) package on the
 #'   three major OSs (linux, macOS and Windows). This action is what the
 #'   tidyverse teams uses on their repositories, but is overkill for less
 #'   widely used packages, which are better off using the simpler
-#'   `use_action_check_release()`.
-#' @rdname actions
+#'   `use_github_action_check_release()`.
+#' @rdname github_actions
 #' @export
-use_action_check_full <- function(save_as = "R.yaml", ignore = TRUE, open = FALSE) {
-  use_action("check-full.yaml", save_as = save_as, ignore = ignore, open = open)
+use_github_action_check_full <- function(save_as = "R-CMD-check.yaml", ignore = TRUE, open = FALSE) {
+  use_github_action("check-full.yaml", save_as = save_as, ignore = ignore, open = open)
 
   use_github_actions_badge("R-CMD-check")
 }
 
-#' @section `use_action_pr_commands()`:
+#' @section `use_github_action_pr_commands()`:
 #' This workflow enables the use of 2 R specific commands in pull request
 #'   issue comments. `\document` will use
 #'   [roxygen2](https://roxygen2.r-lib.org/) to rebuild the documentation for
 #'   the package and commit the result to the pull request. `\style` will use
 #'   [styler](https://styler.r-lib.org/) to restyle your package.
-#' @rdname actions
+#' @rdname github_actions
 #' @export
-use_action_pr_commands <- function(save_as = "pr-commands.yaml", ignore = TRUE, open = FALSE) {
-  use_action("pr-commands.yaml", save_as = save_as, ignore = ignore, open = open)
+use_github_action_pr_commands <- function(save_as = "pr-commands.yaml", ignore = TRUE, open = FALSE) {
+  use_github_action("pr-commands.yaml", save_as = save_as, ignore = ignore, open = open)
+}
+
+uses_github_actions <- function(base_path = proj_get()) {
+  path <- glue("{base_path}/.github/workflows")
+  file_exists(path)
+}
+
+check_uses_github_actions <- function(base_path = proj_get()) {
+  if (uses_github_actions(base_path)) {
+    return(invisible())
+  }
+
+  ui_stop(
+    "
+    Cannot detect that package {ui_value(project_name(base_path))} already uses Github Actions.
+    Do you need to run {ui_code('use_github_actions()')}?
+    "
+  )
 }